Created base models
