--- websites/base/models/__init__.py
@@ -1,11 +1,0 @@
-from .home import HomePage
-from .folder import FolderPage
-from .flex import FlexPage
-from .article import ArticlePage
-from .device import DevicePage
-from .project import ProjectPage
-from .collection import CollectionPage
-from .search import SearchPage
-
-from .index import IndexPage
-from .index_category import IndexCategoryPage
