--- websites/base/models/article.py
@@ -1,7 +1,0 @@
-from abstract.pages.article import AbstractArticlePage
-
-class ArticlePage(AbstractArticlePage):
-    parent_page_types = ["IndexPage"]
-    subpage_type = []
-
-    template = "pages/article.html"
