--- websites/base/models/search.py
@@ -1,8 +1,0 @@
-from abstract.pages.search import AbstractSearchPage
-
-
-class SearchPage(AbstractSearchPage):
-    template = "pages/search.html"
-
-    parent_page_types = ["Homepage"]
-    subpage_type = []
